+++ src/resources/lang/fr/acl.php
@@ -13,4 +13,5 @@
         'group_does_not_exist_with_slug'      => "Il n'y a pas de groupe avec cette limace:",
         'user_does_not_exist_with_name'       => "Il n'y a pas d'utilisateur avec ce nom:",
         'user_does_not_exist'                 => "Il n'y a pas d'utilisateur avec cet identifiant:",
+        'null_model'                          => 'Vous utilisez un modèle null en tant que paramètre.',
     ];